refactor(platform): move identity family out of core
Moves the auth-facing identity values that no `everruns-core` code names into
`everruns-platform`, so they can no longer be imported from `everruns_core`:
OrgMembership, the ANONYMOUS_USER_* seed constants, and the org public-id
generation/validation helpers (generate_org_public_id, validate_org_public_id)
to platform's organization module; PrincipalStatus to platform's principal
module alongside the aggregate it describes. Platform now defines and re-exports
them; core no longer defines or re-exports them.

Kept in core (documented near the code) because core's turn/permissions/runtime
code names or embeds them:
- OrgRole — permissions.rs Caller.role / role_has_permission / Rule::UserHasRole,
  portable tool authorization consumed during a turn.
- DEFAULT_ORG_ID / DEFAULT_ORG_PUBLIC_ID and the internal<->public id conversion
  helpers org_public_id_from_internal / org_internal_id_from_public — used by
  permissions and the knowledge_index capability.
- PrincipalKind + PrincipalSummary — embedded in Session/SessionSchedule/
  AgentIdentity; their serialized JSON and OpenAPI shapes must stay identical.

All consumers (server auth/api/domains/services/seed/tests) import the moved
symbols from everruns_platform. No new dependency edges added to
provider/engine/facade/integrations; the edge stays platform -> core.

Behavior-preserving: serde/OpenAPI shapes unchanged (byte-identical OpenAPI
export), no_platform_dependency and no_platform_store_symbol guards stay green.

Resolves EVE-845.
