+++ crates/server/src/api/schedules.rs
@@ -865,7 +865,8 @@ mod tests {
     use async_trait::async_trait;
     use axum::body::Body;
     use axum::http::Request;
-    use everruns_core::{OrgMembership, OrgRole};
+    use everruns_core::OrgRole;
+    use everruns_platform::OrgMembership;
     use tower::ServiceExt;
 
     use crate::auth::config::{AuthConfig, AuthMode, JwtConfig};
