+++ crates/server/src/auth/middleware.rs
@@ -9,9 +9,12 @@ use axum::{
 };
 use axum_extra::extract::CookieJar;
 use everruns_core::{
-    ANONYMOUS_USER_EMAIL, ANONYMOUS_USER_ID, ANONYMOUS_USER_NAME, Caller, DEFAULT_ORG_ID,
-    DEFAULT_ORG_PUBLIC_ID, DefaultPermissionResolver, FeatureFlags, OrgMembership, OrgRole,
-    PermissionResolver, validate_org_public_id,
+    Caller, DEFAULT_ORG_ID, DEFAULT_ORG_PUBLIC_ID, DefaultPermissionResolver, FeatureFlags,
+    OrgRole, PermissionResolver,
+};
+use everruns_platform::{
+    ANONYMOUS_USER_EMAIL, ANONYMOUS_USER_ID, ANONYMOUS_USER_NAME, OrgMembership,
+    validate_org_public_id,
 };
 use serde::Serialize;
 use std::sync::Arc;
